Commented FoundationStack.cs
FoundationStack.cs is fully commented. I also removed a small redundency in Dr. Cherry's code. Basically, DragOver was calling CanDrop to check which color to highlight the stack, but CanDrop was also highlighting the stack itself instead of just returning a true or false value to DragOver. I removed the highlighting logic from CanDrop so that DragOver is handling the highlighting and CanDrop is just handling the check. I tested it and this didn't break anything.