It now actually keeps count of how many uno reverse cards you have and displays them in the shop. There's also now a class for Uno Reverse cards that will randomly select a color for them and set the image appropriately
I have not yet hooked up the class for the uno reverse cards to the visuals in the shop yet, and those indicators are currently hardcoded as yellow. I'm not sure actually that connecting them would be the way to go with that anyway. I might wind up deleting this class later, idk. I might reuse some of the code from it though when randomizing the color of the cards. idk