It's working!
appearently comparing background images with stored images just isn't something you can do bc when you set a background image winforms converts it to a drawing on the object for performance reasons so the comparison is a drawing vs an image which will never match. I just needed a workaround instead, so I made a parallel array to the array of lightbulbs which is an array of when those lightbulbs are on or off. it's working!